Add files via upload
Some updates:
New Design, this new design is because of all the feedback I've been getting about the addon. I will be adding in a ColorChanger in the future. 
Added all the game modes into 1 button, allowing more space.
Added a second tab, this will be where all the settings / toggles go.
Added Guild % which can be changed, depending on the amount 10% - 90% is best to use.
Added Full Stats button and Fame/Shame (Top 3 winners / Losers) Button.

Tested and confirmed working. 

Added Locolization to addon in next update, looking for people to translate for me. Thanks!
More to come...